[CLEANUP][DB] Misc cleanup changes for DB support - Change all ID usages to Long - Retrieve DbObject from dbId wherever required to fix errors due to changes in Objects - Use dbId instead of old id for Lead - Use SyncDbTask Instead of GetTasks whereever necessary - Add RefreshTasks static API to homescreen and use everywhere - Add static API to FormField to use for converting from JSonArray to Object - Update Homescreen DB Sync and UI Init mechanism - Add Sync button to homescreen - Remove Static APIs which were earlier storing Tasks DB - Add Form Saving & Task closing support to SubmitFormTask. Form should be saved after retrieving id and version from server

+++ app/src/main/java/com/biz/navimate/interfaces/IfaceResult.java
@@ -1,7 +1,5 @@
 package com.biz.navimate.interfaces;
 
-import android.graphics.Bitmap;
-
 import java.util.ArrayList;
 
 /**
@@ -23,7 +21,7 @@ public class IfaceResult {
 
     public interface LeadPicker
     {
-        void onLeadPicked(ArrayList<Integer> leads);
+        void onLeadPicked(ArrayList<Long> leads);
     }
 
     public interface Zxing
